rate limiter

+++ internal/layer/entity/rate_limiter.go
@@ -3,5 +3,5 @@ package entity
 type RateLimiter struct {
 	IP                string `db:"ip"`
 	AllowanceRequests int    `db:"allowance"`
-	Timestamp         int    `db:"timestamp"`
+	Timestamp         int64  `db:"timestamp"`
 }
